use lazy TCP socket
Use a wrapper around TCP socket to talk to the device.

The lazy socket will disconnect after some time of inactivity,
re-connect one we start sending new commands to the device.

This way, we always have a 'fresh' TCP socket. Hopefully,
we'll avoid problems with connection timing out, due to
not sending any packets in a long time.